test
Click "Commit to main"
Click "Publish repository" (if this is the first time) or "Push origin" (if the repository already exists)
This will be much more user-friendly than using command line Git. The GitHub Desktop app will handle all the Git commands for you and show you a visual interface of your changes.